requirements.txt
pandas==1.3.0
openpyxl==3.0.7
streamlit==0.85.0
plotly-express==0.4.1
matplotlib-inline==0.1.2
